refactor(api): move shared api server control methods to 'api/server'
By moving these to a separate package, the other packages can refer to
it, which leaves the original api package only as handlers for the
top-level APIs.  This also allows for those handlers to refer to the
other packages.
