+++ internal/debug/stack.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 	"runtime"
 
-	"github.com/AlexxIT/go2rtc/internal/api"
+	api "github.com/AlexxIT/go2rtc/internal/api/server"
 )
 
 var stackSkip = [][]byte{
