+++ internal/ffmpeg/device/devices.go
@@ -6,7 +6,7 @@ import (
 	"strconv"
 	"sync"
 
-	"github.com/AlexxIT/go2rtc/internal/api"
+	api "github.com/AlexxIT/go2rtc/internal/api/server"
 )
 
 func Init(bin string) {
